rename HelloWorld component
